Fix. Code. AntiCrawler `clearDataTable` fixed.

+++ lib/Cleantalk/ApbctWP/Firewall/AntiCrawler.php
@@ -154,10 +154,10 @@ class AntiCrawler extends \Cleantalk\Common\Firewall\FirewallModule
     {
         $db->execute("TRUNCATE TABLE {$db__table__data};");
         $db->setQuery("SELECT COUNT(*) as cnt FROM {$db__table__data};")->fetch(); // Check if it is clear
-        if ( $db->result['cnt'] != 0 ) {
+        if ( isset($db->result['cnt']) && $db->result['cnt'] != 0 ) {
             $db->execute("DELETE FROM {$db__table__data};"); // Truncate table
             $db->setQuery("SELECT COUNT(*) as cnt FROM {$db__table__data};")->fetch(); // Check if it is clear
-            if ( $db->result['cnt'] != 0 ) {
+            if ( isset($db->result['cnt']) && $db->result['cnt'] != 0 ) {
                 return array('error' => 'COULD_NOT_CLEAR_UA_BL_TABLE'); // throw an error
             }
         }
